stories Mlcenterpos + mlComponentTemplate in TSX
