--- src/components/MlCenterPosition/MlCenterPosition.stories.js
@@ -1,21 +1,0 @@
-import React from "react";
-
-import MlCenterPosition from "./MlCenterPosition";
-
-import mapContextDecorator from "../../decorators/MapContextDecorator";
-
-const storyoptions = {
-	title: "MapComponents/MlCenterPosition",
-	component: MlCenterPosition,
-	argTypes: {
-	},
-	decorators: mapContextDecorator,
-	parameters: { docs: { source: { type: 'code' } } }
-};
-export default storyoptions;
-
-const Template = (args) => <MlCenterPosition />;
-
-export const ExampleConfig = Template.bind({});
-ExampleConfig.parameters = {};
-ExampleConfig.args = {};
